requirements.txt
streamlit==1.41.0
groq==0.12.0
supabase==2.9.0
pymupdf==1.24.0
python-dotenv==1.0.0
requests==2.32.0
